# api.yaml
openapi: 3.0.3
info:
  title: "Подари мне"
  description: API для Telegram Mini-App управления вишлистами и подарками
  version: 1.0.0
  contact:
    name: Development Team
    email: dev@podarimne.ru
servers:
  - url: https://api.podarimne.ru/v1
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Wish:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
          maxLength: 100
        photo:
          type: string
          nullable: true
        url_gift:
          type: string
          maxLength: 2048
          nullable: true
        price:
          type: number
          format: float
          minimum: 0
          maximum: 10000000
          nullable: true
        currency:
          type: string
          enum: [RUB, BYN, USD, EUR, UAH, KZT]
          nullable: true
        description:
          type: string
          maxLength: 500
        status:
          type: boolean
          default: false
        created_at:
          type: string
          format: date-time
        updated_ad:
          type: string
          format: date-time
    Wishlist:
      type: object
      required:
        - id
        - name
        - privacy
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
          maxLength: 50
        description:
          type: string
          maxLength: 250
          nullable: true
        photo:
          type: string
          nullable: true
        tipe_privacy:
          type: string
          enum: [private, limited, public]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /wishes:
    get:
      summary: "Получить список желаний пользователя"
      tags:
        - "Желания"
      security:
        - BearerAuth: []
      responses:
        '200':
          description: "Список желаний"
          content:
            application/json:
              schema:
                type: object
                properties:
                  wishes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Wish'

    post:
      summary: "Создать новое желание"
      tags:
        - "Желания"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  maxLength: 100
                photo:
                  type: string
                url_gift:
                  type: string
                  maxLength: 2048
                price:
                  type: number
                  minimum: 0
                  maximum: 10000000
                currency:
                  type: string
                  enum: [RUB, BYN, USD, EUR, UAH, KZT]
                description:
                  type: string
      responses:
        '201':
          description: "Желание создано"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wish'
        '400':
          description: "Неверные данные"
  /wishes/{wish_id}:
    get:
      summary: "Получить информацию о желании"
      tags:
        - "Желания"
      security:
        - BearerAuth: []
      parameters:
        - name: wish_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: "Информация о желании"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wish'
        '404':
          description: "Желание не найдено"

    put:
      summary: "Обновить желание"
      tags:
        - "Желания"
      security:
        - BearerAuth: []
      parameters:
        - name: wish_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100

                photo:
                  type: string
                url_gift:
                  type: string
                  maxLength: 2048
                price:
                  type: number
                  minimum: 0
                  maximum: 10000000
                currency:
                  type: string
                  enum: [RUB, BYN, USD, EUR, UAH, KZT]
                description:
                  type: string
                  maxLength: 500
                status:
                  type: boolean
      responses:
        '200':
          description: "Желание обновлено"
        '404':
          description: "Желание не найдено"

    delete:
      summary: "Удалить желание"
      tags:
        - "Желания"
      security:
        - BearerAuth: []
      parameters:
        - name: wish_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: "Желание удалено"
        '404':
          description: "Желание не найдено"
  /wishlists:
    get:
      summary: "Получить список вишлистов пользователя"
      tags:
        - "Вишлисты"
      security:
        - BearerAuth: []
      responses:
        '200':
          description: "Список вишлистов"
          content:
            application/json:
              schema:
                type: object
                properties:
                  wishlists:
                    type: array
                    items:
                      $ref: '#/components/schemas/Wishlist'
                  total:
                    type: integer

    post:
      summary: "Создать новый вишлист"
      tags:
        - "Вишлисты"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - privacy
              properties:
                name:
                  type: string
                  maxLength: 50
                description:
                  type: string
                  maxLength: 250
                photo:
                  type: string
                privacy:
                  type: string
                  enum: [private, limited, public]
      responses:
        '201':
          description: "Вишлист создан"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wishlist'
  /wishlists/{wishlist_id}:
    get:
      summary: "Получить информацию о вишлисте"
      tags:
        - "Вишлисты"
      security:
        - BearerAuth: []
      parameters:
        - name: wishlist_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: "Информация о вишлисте"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wishlist'
        '403':
          description: "Нет доступа к вишлисту"
        '404':
          description: "Вишлист не найден"

    put:
      summary: "Обновить вишлист"
      tags:
        - "Вишлисты"
      security:
        - BearerAuth: []
      parameters:
        - name: wishlist_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 50
                description:
                  type: string
                  maxLength: 250
                photo:
                  type: string
                privacy:
                  type: string
                  enum: [private, limited, public]
      responses:
        '200':
          description: "Вишлист обновлен"
        '404':
          description: "Вишлист не найден"

    delete:
      summary: "Удалить вишлист"
      tags:
        - "Вишлисты"
      security:
        - BearerAuth: []
      parameters:
        - name: wishlist_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: "Вишлист удален"
        '404':
          description: "Вишлист не найден"
...
  /wishlists/{wishlist_id}/wishes:
    get:
      summary: "Получить желания в вишлисте"
      tags:
        - "Вишлисты"
      security:
        - BearerAuth: []
      parameters:
        - name: wishlist_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: "Список желаний в вишлисте"
          content:
            application/json:
              schema:
                type: object
                properties:
                  wishes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Wish'
                  total:
                    type: integer

    post:
      summary: "Добавить желание в вишлист"
      tags:
        - "Вишлисты"
      security:
        - BearerAuth: []
      parameters:
        - name: wishlist_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - wish_id
              properties:
                wish_id:
                  type: integer
      responses:
        '200':
          description: "Желание добавлено в вишлист"
        '404':
          description: "Вишлист или желание не найдено"
...